requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-dateutil==2.8.2
fake-useragent==1.4.0
pandas==2.1.4
pyinstaller==6.3.0 